+++ DifyAI.Test/ChatMessagesTest.cs
@@ -15,12 +15,12 @@ namespace DifyAI.Test
         {
             var req = new ChatCompletionRequest
             {
+                ApiKey = "app-Bmqhdl4yuJypTwuIslmdgMyy",
                 Query = "你好",
                 User = "user123",
                 Inputs = new Dictionary<string, string>
                 {
-                    ["arg1"] = "1",
-                    ["arg2"] = "2",
+                    ["Name"] = "Name",
                 },
             };
             var rsp = await _difyAIService.ChatMessages.ChatAsync(req);