+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # 更新历史
 
+## 3.4.0
+- 将 `CompletionMessages.CompletionAsync/CompletionStreamAsync`参数由`ChatCompletionRequest`改为`CompletionRequest` [#14](https://github.com/BitBrewing/dify-csharp-sdk/issues/14)
+
 ## 3.3.0
 
 - 修复: 停止工作流的请求url错误 [#11](https://github.com/BitBrewing/dify-csharp-sdk/issues/11)
 - 修复: 添加缺少的创建知识库的请求参数 [#9](https://github.com/BitBrewing/dify-csharp-sdk/issues/9#issuecomment-2795688431)
 - 修复: 添加缺少的通过文件更新文档的请求参数 [#9](https://github.com/BitBrewing/dify-csharp-sdk/issues/9#issuecomment-2795688431)
-- 新增: 检索知识库接口（DataSets.RetrieveDatasetAsync） [#9](https://github.com/BitBrewing/dify-csharp-sdk/issues/9#issuecomment-2795688431)
-- 修复: 在ProcessRule这个类里，缺失了父分段的召回模式、子分段规则等参数 [#9](https://github.com/BitBrewing/dify-csharp-sdk/issues/9#issuecomment-2795688431)
+- 新增: 检索知识库接口（`DataSets.RetrieveDatasetAsync`） [#9](https://github.com/BitBrewing/dify-csharp-sdk/issues/9#issuecomment-2795688431)
+- 修复: 在`ProcessRule`这个类里，缺失了父分段的召回模式、子分段规则等参数 [#9](https://github.com/BitBrewing/dify-csharp-sdk/issues/9#issuecomment-2795688431)